fix: resolve compilation errors and complete modular service architecture
- Create missing service modules (sample_service.rs, sequencing_service.rs, template_service.rs)
- Fix storage service type mismatch by boxing StorageError in get_file method
- Create missing validation modules (rules.rs, validators.rs) with concrete implementations
- Fix type cast issue in PaginatedResult::new method
- All modules now implement proper service traits with health checks
- Build now passes with only minor warnings (unused imports/variables)

This completes the IKEA-style modular architecture allowing independent
component development and democratic assembly patterns.

+++ src/services/mod.rs
@@ -103,7 +103,7 @@ pub struct PaginatedResult<T> {
 
 impl<T> PaginatedResult<T> {
     pub fn new(items: Vec<T>, total: u64, page: u32, per_page: u32) -> Self {
-        let has_next = (page * per_page) < total;
+        let has_next = ((page * per_page) as u64) < total;
         let has_prev = page > 1;
 
         Self {
